Put second level collapse under inner_hits
